requirements.txt
flask
requests
python-dotenv
sqlalchemy
fake-useragent
beautifulsoup4